Avoid streaming a char16_t to a (narrow) stream.
This overload was removed in C++20.

Bug: 1284275
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3791810
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1029563}

+++ ui/base/ime/linux/input_method_auralinux_unittest.cc
@@ -269,7 +269,7 @@ class TextInputClientForTesting : public DummyTextInputClient {
 
   void InsertChar(const ui::KeyEvent& event) override {
     std::stringstream ss;
-    ss << event.GetCharacter();
+    ss << static_cast<uint16_t>(event.GetCharacter());
     TestResult::GetInstance()->RecordAction(u"keypress:" +
                                             base::ASCIIToUTF16(ss.str()));
   }
